Add files via upload
This is a complete  sample of Voting Machine System in OOP. The primary thing of this voting machine design is
to cast votes, track who is leading candidates, and count the votes. In our voting
machine, before casting a vote, the voter must be registered in the district in which he
wants to vote; he cannot give a vote if he has already cast a vote. In addition to this
project, we will add Admin Login ( with password fast@pk# ). When Admin enters the
district name , he can see all voters' details by the district. He can view voter statistics
based on gender and age. He can distinguish various age groups. He can view voter
statistics based on gender and age.